Revert import integration tests change
