+++ dot/import_test.go
@@ -12,7 +12,6 @@ import (
 
 	"github.com/ChainSafe/gossamer/dot/types"
 	"github.com/ChainSafe/gossamer/lib/common"
-	"github.com/ChainSafe/gossamer/lib/trie"
 	"github.com/ChainSafe/gossamer/lib/utils"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -137,46 +136,6 @@ func Test_newHeaderFromFile(t *testing.T) {
 				assert.NoError(t, err)
 			}
 			assert.Equal(t, tt.want, got)
-		})
-	}
-}
-
-func Test_newTrieFromPairs(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		name     string
-		filename string
-		want     common.Hash
-		err      error
-	}{
-		{
-			name: "no_arguments",
-			err:  errors.New("read .: is a directory"),
-			want: common.Hash{},
-		},
-		{
-			name:     "working example",
-			filename: setupStateFile(t),
-			want:     common.MustHexToHash("0x09f9ca28df0560c2291aa16b56e15e07d1e1927088f51356d522722aa90ca7cb"),
-		},
-	}
-	for _, tt := range tests {
-		tt := tt
-		t.Run(tt.name, func(t *testing.T) {
-			t.Parallel()
-
-			got, err := newTrieFromPairs(tt.filename)
-			if tt.err != nil {
-				assert.EqualError(t, err, tt.err.Error())
-			} else {
-				assert.NoError(t, err)
-			}
-			if tt.want.IsEmpty() {
-				assert.Nil(t, got)
-			} else {
-				assert.Equal(t, tt.want, got.MustHash(trie.NoMaxInlineValueSize))
-			}
 		})
 	}
 }
